Only build once per test run via `pretest`
Whilst running `jest --watch` we were doing a full build on every edit
