+++ jest-puppeteer.config.js
@@ -1,12 +1,9 @@
 const configPaths = require('./lib/paths.js')
 
-const PORT = configPaths.testPort
-
 module.exports = {
   browserContext: 'incognito',
   server: {
-    command: 'node tasks/test-serve.js',
-    launchTimeout: 30000,
-    port: PORT
+    command: 'npm run serve',
+    port: configPaths.testPort
   }
 }
