+++ package.json
@@ -15,9 +15,11 @@
   "scripts": {
     "postinstall": "npm ls --depth=0",
     "build": "node tasks/build.js",
-    "start": "node tasks/serve.js",
+    "start": "node tasks/start.js",
+    "serve": "node tasks/serve.js",
+    "pretest": "npm run build",
     "test": "npm run lint && jest",
-    "lint": "npm run lint:js && npm run lint:scss && npm run lint:html",
+    "lint": "npm run lint:js && npm run lint:scss && npm run lint:html --ignore-scripts",
     "prelint:html": "npm run build",
     "lint:html": "html-validate \"deploy/public/**/index.html\"",
     "lint:js": "eslint --cache --cache-location .cache/eslint --cache-strategy content --color --ignore-path .gitignore \"**/*.{cjs,js,mjs}\"",
